lib: Remove xilmem and xilstandalone library dependencies
Standalone BSP modifications to create separate libraries
xilmem and xilstandalone have been reverted back.
Now standalone BSP code will be part of libxil.a, same as
that of 2018.3 release. Updating libraries and template
applications to remove the dependency on xilmem and
xilstandalone library, to be in line with standalone BSP.

Signed-off-by: Mubin Usman Sayyed <mubin.usman.sayyed@xilinx.com>
Acked-for-series: Anirudha Sarangi <anirudh@xilinx.com>