aiengine: src: dma: tile: Don't check the ready state
Don't check the ready state with XAIE_WRKARND_CRVO1692 as it's not
set at all. Instead it can cause problems, because the tile instance
may not be initialized properly with the statue value set as ready
which skips the initialization. Since most of initialization
other than setting the address is planned to move out, the check can
be removed.

Signed-off-by: Hyun Kwon <hyun.kwon@xilinx.com>

Acked-for-series: Wendy Liang <wendy.liang@xilinx.com>