canfd: Fix build failures in examples
The previous CANFD driver is for PL CANFD controller, hence the CANFD
interrupt id will point to fabric interrupt. But now the driver is
common for both PS and PL and when you import this example for PS CANFD,
then this example is failed to compile.
So fix this by generating macros in xparameters.h with the help of
IS_PL IP parameter and use those macros in interrupt example.

This patch also updates the tcl to generate base addresses of each version
of CANFD controllers.
i.e. For PL CANFD versions, the IP parameter for base address is C_BASEADDR
and for PS CANFD versions, it is C_S_AXI_BASEADDR.

Signed-off-by: Naga Sureshkumar Relli <naga.sureshkumar.relli@xilinx.com>

Acked-for-series: Srinivas Goud <srinivas.goud@xilinx.com>