pmufw: Add check for no of users for vpll
For zynqmp it is assumed that vpll will be used only by dp_video_ref.
Add a check for the same.

Add check only when changing rates. This helps to remove the warn in cases
where the dp_video_ref is not using vpll and vpll is used for other stuff.

Signed-off-by: Rajan Vaja <rajan.vaja@xilinx.com>
Acked-by: Jolly Shah <jollys@xilinx.com>

Acked-by: Jolly Shah <jollys@xilinx.com>