qspips: Read Configuration Registers of both the flashes in dual parellel mode
While reading the configuration registers of the flahses in dual parallel
mode. The qspi controller gives us a 4byte data.The first byte and the
third byte of the received data holds the configuration register values
of the upper and lower flash.

This patch retrieves the configuration register values of both the flashes
form the received data in the driver and passes it to the application.

Signed-off-by: Amit Kumar Mahapatra <amit.kumar-mahapatra@xilinx.com>

Acked-for-series: Naga Sureshkumar Relli <nagasure@xilinx.com>